Feature/tagged stream input (#56)
* Added conda package installation guide

* fix data format between fft_demod and deinterleaver

* added support of tagged stream input

* add frequency issue related to repo update

* change uint to uint32_t for windows build

* removed utilities.h from python bindings
Prevents issue with out of date bindings when editing utilities.h

* add comments and documentation

+++ lib/crc_verif_impl.cc
@@ -104,7 +104,7 @@ namespace gr
             consume_each(ninput_items[0]);
 
 
-            if ((in_buff.size() >= (int)m_payload_len + 2) && m_crc_presence)
+            if ((in_buff.size() >= m_payload_len + 2) && m_crc_presence)
             { // wait for all the payload to come
 
                 if (m_payload_len < 2)
@@ -174,7 +174,7 @@ namespace gr
                         return m_payload_len;
                 }
             }
-            else if ((in_buff.size()>= (int)m_payload_len) && !m_crc_presence)
+            else if ((in_buff.size()>= m_payload_len) && !m_crc_presence)
             {
                 if (output_items.size()){
 		            curent_tag.offset = nitems_written(0);
