+++ lib/frame_sync_impl.cc
@@ -138,7 +138,7 @@ namespace gr
             // Dechirping
             volk_32fc_x2_multiply_32fc(&dechirped[0], samples, &downchirp_aug[0], up_symb_to_use * m_number_of_bins);
             // prepare FFT
-            for (int i = 0; i < 2 * up_symb_to_use * m_number_of_bins; i++)
+            for (uint32_t i = 0; i < 2 * up_symb_to_use * m_number_of_bins; i++)
             {
                 if (i < up_symb_to_use * m_number_of_bins)
                 {
@@ -177,7 +177,7 @@ namespace gr
             k_residual = fmod((k0 + ka) / 2 / up_symb_to_use, 1);
             cfo_frac = k_residual - (k_residual > 0.5 ? 1 : 0);
             // Correct CFO frac in preamble
-            for (int n = 0; n < up_symb_to_use * m_number_of_bins; n++)
+            for (uint32_t n = 0; n < up_symb_to_use * m_number_of_bins; n++)
             {
                 CFO_frac_correc_aug[n] = gr_expj(-2 * M_PI * (cfo_frac) / m_number_of_bins * n);
             }
@@ -208,7 +208,7 @@ namespace gr
                 // Dechirping
                 volk_32fc_x2_multiply_32fc(&dechirped[0], &samples[m_number_of_bins * i], &m_downchirp[0], m_number_of_bins);
                 // prepare FFT
-                for (int j = 0; j < m_number_of_bins; j++)
+                for (uint32_t j = 0; j < m_number_of_bins; j++)
                 {
                     cx_in_cfo[j].r = dechirped[j].real();
                     cx_in_cfo[j].i = dechirped[j].imag();
@@ -238,7 +238,7 @@ namespace gr
             }
             cfo_frac = -std::arg(four_cum) / 2 / M_PI;
             // Correct CFO in preamble
-            for (int n = 0; n < up_symb_to_use * m_number_of_bins; n++)
+            for (uint32_t n = 0; n < up_symb_to_use * m_number_of_bins; n++)
             {
                 CFO_frac_correc_aug[n] = gr_expj(-2 * M_PI * cfo_frac / m_number_of_bins * n);
             }
@@ -269,7 +269,7 @@ namespace gr
                 volk_32fc_x2_multiply_32fc(&dechirped[0], &preamble_upchirps[m_number_of_bins * i], &m_downchirp[0], m_number_of_bins);
 
                 // prepare FFT
-                for (int j = 0; j < 2 * m_number_of_bins; j++)
+                for (uint32_t j = 0; j < 2 * m_number_of_bins; j++)
                 {
                     if (j < m_number_of_bins)
                     {
@@ -325,7 +325,7 @@ namespace gr
             // Multiply with ideal downchirp
             volk_32fc_x2_multiply_32fc(&dechirped[0], samples, ref_chirp, m_number_of_bins);
 
-            for (int i = 0; i < m_number_of_bins; i++)
+            for (uint32_t i = 0; i < m_number_of_bins; i++)
             {
                 cx_in[i].r = dechirped[i].real();
                 cx_in[i].i = dechirped[i].imag();
@@ -364,7 +364,7 @@ namespace gr
             // Multiply with ideal downchirp
             volk_32fc_x2_multiply_32fc(&dechirped[0], samples, &m_downchirp[0], m_number_of_bins);
 
-            for (int i = 0; i < m_number_of_bins; i++)
+            for (uint32_t i = 0; i < m_number_of_bins; i++)
             {
                 cx_in[i].r = dechirped[i].real();
                 cx_in[i].i = dechirped[i].imag();
@@ -457,7 +457,7 @@ namespace gr
             int items_to_output = 0;
 
             // check if there is enough space in the output buffer
-            if (noutput_items < m_number_of_bins)
+            if ((uint32_t)noutput_items < m_number_of_bins)
             {
                 return 0;
             }
@@ -494,7 +494,7 @@ namespace gr
             }
 
             // downsampling
-            for (int ii = 0; ii < m_number_of_bins; ii++)
+            for (uint32_t ii = 0; ii < m_number_of_bins; ii++)
                 in_down[ii] = in[(int)(m_os_factor / 2 + m_os_factor * ii - my_roundf(m_sto_frac * m_os_factor))];
 
             switch (m_state)
@@ -547,7 +547,7 @@ namespace gr
                     m_cfo_frac = estimate_CFO_frac_Bernier(&preamble_raw[m_number_of_bins - k_hat]);
                     m_sto_frac = estimate_STO_frac();
                     // create correction vector
-                    for (int n = 0; n < m_number_of_bins; n++)
+                    for (uint32_t n = 0; n < m_number_of_bins; n++)
                     {
                         CFO_frac_correc[n] = gr_expj(-2 * M_PI * m_cfo_frac / m_number_of_bins * n);
                     }
@@ -562,7 +562,7 @@ namespace gr
                 {
                 case NET_ID1:
                 {
-                    if (bin_idx == 0 || bin_idx == 1 || bin_idx == m_number_of_bins - 1)
+                    if (bin_idx == 0 || bin_idx == 1 || (uint32_t)bin_idx == m_number_of_bins - 1)
                     { // look for additional upchirps. Won't work if network identifier 1 equals 2^sf-1, 0 or 1!
                         memcpy(&net_id_samp[0], &in[(int)0.75 * m_samples_per_symbol], sizeof(gr_complex) * 0.25 * m_samples_per_symbol);
                         if (additional_upchirps >= 3)
@@ -609,7 +609,7 @@ namespace gr
                 case QUARTER_DOWN:
                 {
                     memcpy(&additional_symbol_samp[m_samples_per_symbol], &in[0], sizeof(gr_complex) * m_samples_per_symbol);
-                    if (down_val < m_number_of_bins / 2)
+                    if ((uint32_t)down_val < m_number_of_bins / 2)
                     {
                         m_cfo_int = floor(down_val / 2);
                     }
@@ -623,7 +623,7 @@ namespace gr
 
                     std::vector<gr_complex> CFO_int_correc;
                     CFO_int_correc.resize((m_n_up_req + additional_upchirps) * m_number_of_bins);
-                    for (int n = 0; n < (m_n_up_req + additional_upchirps) * m_number_of_bins; n++)
+                    for (uint32_t n = 0; n < (m_n_up_req + additional_upchirps) * m_number_of_bins; n++)
                     {
                         CFO_int_correc[n] = gr_expj(-2 * M_PI * (m_cfo_int) / m_number_of_bins * n);
                     }
@@ -639,7 +639,7 @@ namespace gr
                     int N = m_number_of_bins;
                     std::vector<gr_complex> sfo_corr_vect;
                     sfo_corr_vect.resize((m_n_up_req + additional_upchirps) * m_number_of_bins, 0);
-                    for (int n = 0; n < (m_n_up_req + additional_upchirps) * m_number_of_bins; n++)
+                    for (uint32_t n = 0; n < (m_n_up_req + additional_upchirps) * m_number_of_bins; n++)
                     {
                         sfo_corr_vect[n] = gr_expj(-2 * M_PI * (pow(mod(n, N), 2) / 2 / N * (m_bw / fs_p * m_bw / fs_p - m_bw / fs * m_bw / fs) + (std::floor((float)n / N) * (m_bw / fs_p * m_bw / fs_p - m_bw / fs_p) + m_bw / 2 * (1 / fs - 1 / fs_p)) * mod(n, N)));
                     }
@@ -657,7 +657,7 @@ namespace gr
                     std::vector<gr_complex> corr_preamb;
                     corr_preamb.resize((m_n_up_req + additional_upchirps) * m_number_of_bins, 0);
                     // apply sto correction
-                    for (int i = 0; i < (m_n_up_req + additional_upchirps) * m_number_of_bins; i++)
+                    for (uint32_t i = 0; i < (m_n_up_req + additional_upchirps) * m_number_of_bins; i++)
                     {
                         corr_preamb[i] = preamble_raw_up[m_os_factor * (m_number_of_bins - k_hat + i) - int(my_roundf(m_os_factor * m_sto_frac))];
                     }
@@ -691,7 +691,7 @@ namespace gr
                     net_ids_samp_dec.resize(2 * m_number_of_bins, 0);
                     // start_off gives the offset in the net_id_samp vector required to be aligned in time (CFOint is equivalent to STOint since upchirp_val was forced to 0)
                     int start_off = (int)m_os_factor / 2 - (my_roundf(m_sto_frac * m_os_factor)) + m_os_factor * (.25 * m_number_of_bins + m_cfo_int);
-                    for (int i = 0; i < m_number_of_bins * 2; i++)
+                    for (uint32_t i = 0; i < m_number_of_bins * 2; i++)
                     {
                         net_ids_samp_dec[i] = net_id_samp[start_off + i * m_os_factor];
                     }
@@ -833,7 +833,7 @@ namespace gr
             {
                 // transmit only useful symbols (at least 8 symbol for PHY header)
 
-                if (symbol_cnt < 8 || (symbol_cnt < m_symb_numb && m_received_head))
+                if (symbol_cnt < 8 || ((uint32_t)symbol_cnt < m_symb_numb && m_received_head))
                 {
                     // output downsampled signal (with no STO but with CFO)
                     memcpy(&out[0], &in_down[0], m_number_of_bins * sizeof(gr_complex));
